Refactors transport layer.
This extracts/isolates the transport layer from the rest of the code in ethrpc.  This gives a layer of abstraction that allows each transport type to deal with queueing, reconnecting and errors in isolation without the rest of the code needing to have any understanding about them, their differences or transient connectivity issues.

On connect, one or more will be instantiated and connections will attempted to be established internally.  If at least one transport is working, then the application will proceed.  If no transports are able to be established then connect will fail.

When a request is made, if it fails in a way that is retryable then the transport will retry under-the-hood, reconnecting if necessary, only calling back to the user know once the request has gone through.