Creates source directory and moves files into it.
In the process of doing this, I found that the linter wasn't actually linting the entire project.  Upon fixing that, I find several linting errors (some of which caught real bugs) which are also fixed in this PR.

I disabled `laxbreak` linter error partly because it is deprecated: http://jshint.com/docs/options/#laxbreak and partly because of the reasons outlined in [this comment](http://stackoverflow.com/questions/15140740/explanation-of-jshints-bad-line-breaking-before-error#comment39683849_15141168) and [this comment](http://stackoverflow.com/questions/15140740/explanation-of-jshints-bad-line-breaking-before-error#comment41169545_15140740).  Research suggests the reason it originally made it into JSHint (which tries to be un-opinionated) is because there was FUD around the automatic semicolon inserter, but it turns out this linting option doesn't address that at all, which is why I believe it is being deprecated.

I have verified that all of the gulp and NPM tasks still work correctly (which includes all tests passing).