Adds chain management.
This enables ethrpc to manage the blockchain client-side, rather than relying on the Ethereum node to do block chain management.  This gives much stronger guarantees that during a reorg or period of connectivity loss you will get notifications for all added and removed blocks.  This is a precursor to also getting strong guarantees of well ordered log message additions/removals.