Skip transport disconnect/reconnect on retryable error with HttpTransport
https://app.clubhouse.io/augur/story/17842/ethrpc-don-t-issue-disconnect-on-http-retry-unless-socket-error

Currently, during an ethrpc request, if HttpTransport gets a retryable error,
it'll first disconnect/reconnect the transport before retrying. For the error
"0x" result, which indicates data unavailability, we don't want to call
disconnect, because downstream consumers might interpret the disconnect as being
related to a non-retryable issue.

This change adds a `skipReconnect` flag to skip this disconnect/reconnect.
`skipReconnect` is only used on "0x" error in HttpTransport.

Testing done: added local log messages to verify that disconnect/reconnect was
skipped if and only if skipReconnect flag was set.